# Remove check for primitive/builtin in `as_mapper.default()` to improve performance