--- a/R/map-mapper.R
+++ b/R/map-mapper.R
@@ -43,18 +43,7 @@ as_mapper <- function(.f, ...) {
 
 #' @export
 as_mapper.default <- function(.f, ...) {
-  if (typeof(.f) %in% c("special", "builtin")) {
-    .f <- rlang::as_closure(.f)
-
-    # Workaround until fixed in rlang
-    if (is_reference(fn_env(.f), base_env())) {
-      environment(.f) <- global_env()
-    }
-
-    .f
-  } else {
     rlang::as_function(.f)
-  }
 }
 
 #' @export