@hadley 2025-09-19
Could you please add a news bullet too?